+++ src/django/api/serializers/v1/production_locations_serializer.py
@@ -1,36 +1,62 @@
-from rest_framework import serializers
+from rest_framework.serializers import (
+    CharField,
+    ChoiceField,
+    FloatField,
+    IntegerField,
+    ListField,
+    Serializer,
+    ValidationError
+)
 from api.serializers.v1.opensearch_error_list_builder  \
     import OpenSearchErrorListBuilder
+from api.serializers.v1.opensearch_common_validators. \
+    countries_validator import CountryValidator
 from api.serializers.v1.opensearch_common_validators. \
     number_of_workers_validator import NumberOfWorkersValidator
 from api.serializers.v1.opensearch_common_validators. \
     percent_of_female_workers import PercentOfFemaleWorkersValidator
 from api.serializers.v1.opensearch_common_validators. \
     coordinates_validator import CoordinatesValidator
+from api.views.v1.utils import COMMON_ERROR_MESSAGE
 
 
-class ProductionLocationsSerializer(serializers.Serializer):
-
-    number_of_workers_min = serializers.IntegerField(required=False)
-    number_of_workers_max = serializers.IntegerField(required=False)
-    percent_female_workers_min = serializers.FloatField(required=False)
-    percent_female_workers_max = serializers.FloatField(required=False)
-    coordinates_lat = serializers.FloatField(required=False)
-    coordinates_lon = serializers.FloatField(required=False)
+class ProductionLocationsSerializer(Serializer):
+    # These params are checking considering serialize_params output
+    size = IntegerField(required=False)
+    number_of_workers_min = IntegerField(required=False)
+    number_of_workers_max = IntegerField(required=False)
+    percent_female_workers_min = FloatField(required=False)
+    percent_female_workers_max = FloatField(required=False)
+    coordinates_lat = FloatField(required=False)
+    coordinates_lon = FloatField(required=False)
+    country = ListField(
+        child=CharField(required=False),
+        required=False
+    )
+    sort_by = ChoiceField(
+        choices=['name', 'address'],
+        required=False
+    )
+    order_by = ChoiceField(
+        choices=['asc', 'desc'],
+        required=False
+    )
 
     def validate(self, data):
         validators = [
             NumberOfWorkersValidator(),
             PercentOfFemaleWorkersValidator(),
-            CoordinatesValidator()
+            CoordinatesValidator(),
+            CountryValidator(),
         ]
 
         error_list_builder = OpenSearchErrorListBuilder(validators)
         errors = error_list_builder.build_error_list(data)
 
         if errors:
-            raise serializers.ValidationError({
-                "message": "The request query is invalid.",
+            # TODO: Pass error msg to the Rollbar here
+            raise ValidationError({
+                "message": COMMON_ERROR_MESSAGE,
                 "errors": errors
             })