+++ src/django/api/views/v1/utils.py
@@ -1,33 +1,66 @@
+import logging
+from functools import wraps
 from rest_framework.response import Response
 from rest_framework import status
-import logging
+from api.views.v1.parameters_list import V1_PARAMETERS_LIST
+from api.services.search import OpenSearchServiceException
+
 logger = logging.getLogger(__name__)
 
 
+COMMON_ERROR_MESSAGE = 'The request query is invalid.'
+
+
 def serialize_params(serializer_class, query_params):
-    flattened_query_params = {
-        key.replace(']', '').replace('[', '_'): value
-        for key, value in query_params.items()
-    }
+    flattened_query_params = {}
+    for key, value in query_params.lists():
+        # Convert deepObject params
+        if key in [
+            f'{V1_PARAMETERS_LIST.NUMBER_OF_WORKERS}[min]',
+            f'{V1_PARAMETERS_LIST.NUMBER_OF_WORKERS}[max]',
+            f'{V1_PARAMETERS_LIST.PERCENT_FEMALE_WORKERS}[min]',
+            f'{V1_PARAMETERS_LIST.PERCENT_FEMALE_WORKERS}[max]',
+            f'{V1_PARAMETERS_LIST.COORDINATES}[lat]',
+            f'{V1_PARAMETERS_LIST.COORDINATES}[lon]',
+        ]:
+            new_key = key.replace(']', '').replace('[', '_')
+            flattened_query_params[new_key] = value[0]
+        # Prepare only single params
+        elif key in [
+            V1_PARAMETERS_LIST.ADDRESS,
+            V1_PARAMETERS_LIST.DESCRIPTION,
+            V1_PARAMETERS_LIST.SEARCH_AFTER,
+            V1_PARAMETERS_LIST.SORT_BY,
+            V1_PARAMETERS_LIST.ORDER_BY,
+            V1_PARAMETERS_LIST.SIZE
+        ]:
+            flattened_query_params[key] = value[0]
+        else:
+            flattened_query_params[key] = value
 
     params = serializer_class(data=flattened_query_params)
+
     if not params.is_valid():
-        error_response = {
-            'message': None,
-            'errors': []
-        }
+        error_response = {'message': None, 'errors': []}
+        # Handle common validation errors.
+        if 'message' not in params.errors and 'errors' not in params.errors:
+            error_response['message'] = COMMON_ERROR_MESSAGE
+            for field, error_list in params.errors.items():
+                error_response['errors'].append({
+                    'field': field,
+                    'message': error_list[0].title()
+                })
 
-        error_response['message'] = (
-            params.errors.get('message')[0].title()
-        )
-
-        for error_item in params.errors.get('errors', []):
-            field = str(error_item.get('field', '')).title()
-            message = str(error_item.get('message', '')).title()
-            error_response['errors'].append({
-                'field': field,
-                'message': message
-            })
+        # Handle errors that come from serializers
+        message_errors = params.errors.get('message')
+        if message_errors:
+            error_response['message'] = message_errors[0].title()
+        if 'message' in params.errors and 'errors' in params.errors:
+            for error_item in params.errors.get('errors', []):
+                error_response['errors'].append({
+                    'field': error_item.get('field', '').title(),
+                    'message': error_item.get('message', '').title()
+                })
 
         return None, error_response
 
@@ -38,7 +71,7 @@ def handle_value_error(e):
     logger.error(f'Error processing request: {e}')
     return Response(
         {
-            "message": "The request query is invalid.",
+            "message": COMMON_ERROR_MESSAGE,
             "errors": [
                 {
                     "field": "general",
@@ -61,3 +94,15 @@ def handle_opensearch_exception(e):
         },
         status=status.HTTP_500_INTERNAL_SERVER_ERROR
     )
+
+
+def handle_errors_decorator(view_func):
+    @wraps(view_func)
+    def _wrapped_view(self, request, *args, **kwargs):
+        try:
+            return view_func(self, request, *args, **kwargs)
+        except ValueError as e:
+            return handle_value_error(e)
+        except OpenSearchServiceException as e:
+            return handle_opensearch_exception(e)
+    return _wrapped_view